refactor: separate aob and program into cargo workspace
