+++ Cargo.toml
@@ -1,0 +1,5 @@
+[workspace]
+members = [
+    "aob",
+    "program"
+]
