+++ aob/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "aob"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.24"
+solana-program = "=1.9.6"
+bytemuck = { version = "1.7.3", features = ["derive", "extern_crate_std"] }
+num_enum = "0.5.4"
+borsh = "0.9.3"
+num-traits = "0.2.14"
+num-derive = "0.3.3"
+enumflags2 = "0.7.1"
+spl-token = { version = "3.2.0", features = ["no-entrypoint"] }
+bonfida-utils = "0.2.2"
+
+[dev-dependencies]
+hexdump = "0.1.0"
+rand = "0.8.4"
+
+
