+++ aob/src/critbit.rs
@@ -2,15 +2,15 @@ use std::convert::identity;
 use std::convert::TryInto;
 
 use borsh::{BorshDeserialize, BorshSerialize};
-use bytemuck::{Pod, try_from_bytes, try_from_bytes_mut, Zeroable};
+use bytemuck::{try_from_bytes, try_from_bytes_mut, Pod, Zeroable};
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
 use solana_program::account_info::AccountInfo;
 use solana_program::pubkey::Pubkey;
 
-use crate::error::{AoError, AoResult};
 use crate::error::AoError::FailedToDeserializeSlabHeader;
-use crate::state::{AccountTag};
+use crate::error::{AoError, AoResult};
+use crate::state::AccountTag;
 
 // A Slab contains the data for a slab header and an array of nodes of a critbit tree
 // whose leafs contain the data referencing an order of the orderbook.
@@ -227,7 +227,7 @@ impl<'a> Slab<'a> {
             .unwrap()
     }
 
-    pub(crate) fn initialize(
+    pub fn initialize(
         bids_account: &AccountInfo<'a>,
         asks_account: &AccountInfo<'a>,
         market_address: Pubkey,
@@ -914,7 +914,7 @@ mod tests {
 
     use super::*;
 
-// #[test]
+    // #[test]
     // fn test_node_serialization() {
     //     let mut rng = StdRng::seed_from_u64(42);
     //     let mut bytes = [0u8; 100];
