+++ aob/src/lib.rs
@@ -1,0 +1,9 @@
+#[doc(hidden)]
+pub mod critbit;
+#[doc(hidden)]
+pub mod error;
+pub mod orderbook;
+pub mod params;
+/// Describes the different data structres that the program uses to encode state
+pub mod state;
+pub mod utils;
