+++ aob/src/utils.rs
@@ -19,18 +19,14 @@ unsafe fn invariant(check: bool) {
 }
 
 // Safety verification functions
-pub(crate) fn check_account_key(
-    account: &AccountInfo,
-    key: &[u8],
-    error: AoError,
-) -> Result<(), AoError> {
+pub fn check_account_key(account: &AccountInfo, key: &[u8], error: AoError) -> Result<(), AoError> {
     if account.key.to_bytes() != key {
         return Err(error);
     }
     Ok(())
 }
 
-pub(crate) fn check_account_owner(
+pub fn check_account_owner(
     account: &AccountInfo,
     owner: &[u8],
     error: AoError,
@@ -41,14 +37,14 @@ pub(crate) fn check_account_owner(
     Ok(())
 }
 
-pub(crate) fn check_signer(account: &AccountInfo) -> ProgramResult {
+pub fn check_signer(account: &AccountInfo) -> ProgramResult {
     if !(account.is_signer) {
         return Err(ProgramError::MissingRequiredSignature);
     }
     Ok(())
 }
 
-pub(crate) fn check_unitialized(account: &AccountInfo) -> AoResult {
+pub fn check_unitialized(account: &AccountInfo) -> AoResult {
     if account.data.borrow()[0] != 0 {
         return Err(AoError::AlreadyInitialized);
     }
@@ -74,12 +70,12 @@ pub fn get_spread<'a>(
 }
 
 /// a is fp0, b is fp32 and result is a/b fp0
-pub(crate) fn fp32_div(a: u64, b_fp32: u64) -> u64 {
+pub fn fp32_div(a: u64, b_fp32: u64) -> u64 {
     (((a as u128) << 32) / (b_fp32 as u128)) as u64
 }
 
 /// a is fp0, b is fp32 and result is a*b fp0
-pub(crate) fn fp32_mul(a: u64, b_fp32: u64) -> u64 {
+pub fn fp32_mul(a: u64, b_fp32: u64) -> u64 {
     (((a as u128) * (b_fp32 as u128)) >> 32) as u64
 }
