+++ program/Cargo.toml
@@ -3,8 +3,6 @@ name = "agnostic-orderbook"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [features]
 no-entrypoint = []
 test-bpf = []
@@ -14,30 +12,24 @@ lib = []
 utils = []
 
 [dependencies]
-solana-program = "=1.9.4"
-bytemuck = { version = "1.7.2", features = ["derive", "extern_crate_std"] }
-num_enum = "0.5.4"
-borsh = "0.9.1"
-thiserror = "1.0.24"
-num-traits = "0.2"
-num-derive = "0.3"
-enumflags2 = "0.7.1"
+aob = { path = "../aob" }
+solana-program = "=1.9.6"
+borsh = "0.9.3"
 spl-token = { version = "3.2.0", features = ["no-entrypoint"] }
-bonfida-utils = "0.2"
+bonfida-utils = "0.2.2"
+num-traits = "0.2.14"
+num-derive = "0.3.3"
 
 [dev-dependencies]
-solana-sdk = "=1.9.4"
-solana-client = "=1.9.4"
-solana-logger = "=1.9.4"
-solana-validator = "=1.9.4"
-solana-program-test = "=1.9.4"
-hexdump = "0.1.0"
-rand = "0.8.4"
+solana-sdk = "=1.9.6"
+solana-client = "=1.9.6"
+solana-logger = "=1.9.6"
+solana-validator = "=1.9.6"
+solana-program-test = "=1.9.6"
 arrayref = "0.3.6"
 tokio = { version = "1.16", features = ["macros"] }
 anyhow = "1.0.53"
-
-
+bytemuck = { version = "1.7.3", features = ["derive", "extern_crate_std"] }
 
 [lib]
 crate-type = ["cdylib", "lib"]
