+++ program/src/lib.rs
@@ -43,24 +43,17 @@ Once event processing is over, it is essential to pop the processed events off t
 primitive. In general, the event processing logic should be handled by a dedicated cranker on the caller program's side.
 */
 
-#[doc(hidden)]
-pub mod entrypoint;
-/// Program instructions and their CPI-compatible bindings
-pub mod instruction;
-/// Describes the different data structres that the program uses to encode state
-pub mod state;
-
-#[doc(hidden)]
-pub mod critbit;
-#[doc(hidden)]
-pub mod error;
-
-pub use processor::msrm_token;
 use solana_program::declare_id;
 
-pub(crate) mod orderbook;
+pub use processor::msrm_token;
+
+#[doc(hidden)]
+pub mod entrypoint;
+
+/// Program instructions and their CPI-compatible bindings
+pub mod instruction;
+
 #[doc(hidden)]
 pub(crate) mod processor;
-pub mod utils;
 
 declare_id!("aaobKniTtDGvCZces7GH5UReLYP671bBkB96ahr9x3e");
